requirements.txt
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
cohere>=4.0.0
mistralai>=0.1.0
groq>=0.4.0
together>=0.2.0
replicate>=0.15.0
huggingface-hub>=0.17.0
python-dotenv>=1.0.0
requests>=2.31.0